requirements.txt
pyspark==3.5.1
python-dotenv==1.0.1
black==24.10.0
